requirements.txt
# Core dependencies
numpy>=1.21.0
matplotlib>=3.4.0

# For data loading
keras>=2.7.0

# Logging and experiment tracking
wandb>=0.12.0
python-dotenv>=1.0.0

# For confusion matrix and metrics calculation
scikit-learn>=0.24.2